Abstract out 'ModelStub' to be shared
